Move RequirementsTxtRequirement to its own module
Small refactoring, no functional changes.
